backend/requirements.txt
# backend/requirements.txt
Flask
Flask-Cors
Werkzeug
mysql-connector-python
pdfplumber
python-docx
spacy